feat: Calculate approximate transcripts for headers. (#19)
Includes a couple of snapshot tests where we serialize the preprocessor
history.

For now, we only set up approximate calculation so that we can focus
on working on the bulk of the indexer (which is more important), rather
than getting the transcript calculation 100% right. We can fix that
later; it is tracked in the follow-up issue #21.