feat: Finish pipeline until index.scip emission (#39)
Workers generate partial indexes, which are merged by the driver.
This patch is large because it implements the full merging logic
for merging partial indexes. Ideally, this merging logic would live
as a portable C/C++ library in the SCIP repo, but leaving it in
here for convenience.

We also bump the C++ version to C++20 for using <=> and newer
string_view APIs.