fix: Fix bug due to mixup between absolute & relative paths.
We should consistently perform lookups in the planner
using absolute paths as certain headers may not be inside
the project. However, we were incorrectly doing lookups
using relative paths. So propagate the project root to
the planner, and use that to convert the relative path
to an absolute one before performing the lookup.