refactor: Properly track compdb index
Some compilation database jobs can be skipped, for example,
if the main file is not present on disk. In this case,
we need to track the index at the time of validation.

Previously, we tracked the index after some elements were
potentially skipped, which would lead to incorrect indexes
being printed in logs/error messages.