fix: Strip common arch-specific flags before processing (#488)
Generalizes the argument dropping mechanism used specifically
for nvcc to also work for Clang and GCC, and skips some
known arch-specific flags for Clang and GCC, to reduce risk
of unknown flags hit in practice on other architectures.

Fixes GRAPH-1187. See comments for more context.

I've not added a full-blown mechanism to skip arbitrary flags
(e.g. via an extra flag), but we could do that later if necessary.

Added some basic unit tests at the level of the command-line
tweaking. It's a bit more difficult to test this logic at a
higher-level, because the layer above it does I/O for compiler
detection.